CDPD-16052 CDPD-16161: Adding missing operators of "%" and "!="
This patch changes the mapping from the Calcite function name to the
corresponding Impala function for the modulo operator, i.e., "%".
Specifically, the field of 'fnName' for the modulo operator is changed
to "%" from "mod".

Moreover, additional 'fnName' to 'impalaFnName' pairs for the not equal
operator are added so that after this patch, the operator of "!=" in
addition to "<>" is supported.

Testing:
 - Verified in a local development environment that the operators of "%"
   and "!=" are supported.

Change-Id: Ic658f49ac06acf8f2106ec082d0ad5ef99f7ce5a