CDPD-16209: Fix creation and propagation of runtime filters
Certain runtime filters were either not created at all or not
propagated past certain operators in the case of TPC-DS q5 and
q95. This adversely affected performance.  Fixed it by:
 - allowing creation of auxiliary predicates in analyzer's
   createAuxEqPredicate().
 - registering the join conjuncts with value transfer graph.
 - registering the union operator's slots with value transfer graph.

Verified that q5 and q95 had the expected runtime filters.
The plans for q38 and q87 changed slightly as 1 exchange (between
aggregate operators) was dropped likely because the cardinality
estimate does not change.

Change-Id: I03c8099e62ec5b6514b25a9d45eb91105590b970