CDPD-16234: Add table level hostInfo to support file-metadata
In case of HdfsTables the file-metadata also includes the
block-locations. These block locations include Host network addresses as well.
The network addresses are limited and instead of a network address multiple
times for various files, the unique Host network addresses are sent in a
Dictionary which is at the table level. The file descriptors include only the
integer identifiers of the host network address from the dictionary.

This patch makes changes to the HMS API in support to send these deduplicated
Host network address when requesting file-metadata for a table or partitions.

Note that some of the HMS API changes have not been upstreamed yet. I will
open a follow-up JIRA for that. This is two part change, the second part of the
change will be done in Impala which uses the changed APIs in Hive.

Testing done:
1. Added a new test in Impala code base which requests file-metadata
for a partitioned and unpartitioned hdfs table, deserializes it and compares
it against what is present in the Catalogd cache.

Change-Id: Ice251cb497c68d8fd1c96a7285863e52d81fb91c