CDPD-16526: Refactor Hive Partition pruning code so it is extensible
This is a precursor step to checking in CDPD-16524. This checkin
will allow the Impala partition pruning code to be extended seamlessly.

Two new interfaces are added here.

- PartitionPruneRuleHelper: This interface allows an entry point for
dealing with the partition pruning for a specific engine. An object
implementing this interface will be created at RelOptHiveTable
construction.
- PartitionPredicates: This interface is created from the helper. This
class is necessary so that an immutable object can be created with
objects created when the prune partition rule gets fired.

The partition pruning fetch is called from a couple of places other
than the partition pruning rule. In these cases, all the partitions
need to be fetched (e.g. from getRowCount()). There is no need for
engine specific code to be called here since there is no expression
to be parsed, so it will always use the HivePartitionPruneRuleHelper.

For compilation and testing purposes, the ImpalaFunctionHelper calls
the Hive code. This is temporary and will be changed when CDPD-16524
gets checked in.

Change-Id: I765556f2230c206bfd1a98014a4083ae8235377a