CDPD-16524: Implement partition pruning using Impala semantics.
Impala semantics are now used when pruning the partitions.

At optimization time, the partition names are fetched from HMS.
The PartitionPruner in the Impala code base is used to evaluate
the filter condition expressions and places the partitions to
keep in the ImpalaPrunedPartitionList object.

At translation time, the ImpalaHdsfScanRel object calls into
the ImpalaLightHdfsTable object and the fetch for the partition
and file metadata from HMS is done.

Because the fetch is done at translation time, only one fetch
from HMS will be done even if there are multiple table references
to the same table within the query.

This commit does not contain secrets

Change-Id: I5b94e20d8fa6c3e0db90d8aec4488495e7710aaf