CDPD-16119: Comparison of literal string and integer fails with assert.
When running the query "select '1' = 1", the code checks to see if
there is a common datatype for which both datatypes can be cast to.

The impala_scalars.json file allowed both of these types to be cast
to timestamp. This implicit cast should not be allowed and is now removed.

However, on testing, it seems that the following query failed with assert:
select cast(1 as timestamp) = cast('1' as timestamp)

This query failed because all return types for all functions were being
set to 'not null'. This is not the desired behavior.

With this push, it will now be assumed that all functions can return null.
Also, a cast function will always return whatever the arg datatype is, with
one exception.  In the case of timestamp, the "cast('1' as timestamp)" returns
null, so we allow nulls to be returned when the cast is of type 'timestamp'.

Change-Id: I05769b2a396cfd34037678dfdfb2cc75f900074d