CDPD-16663: Call analyze on builtin function names to classify correctly
FunctionName for Builtin functions such as casttobigint() needs
to be analyzed in order to classify it as a builtin even though
the resource file has binaryType: "BUILTIN". This allows it to
be treated the same way as a CAST expr for selectivity estimation.

This patch makes this change for both scalar functions and
aggregate functions. However, aggregate functions also have a
separate path where they call Impala's
AggregateFunction.createBuiltin() directly, so the net effect
of this patch is mainly for builtin scalar functions only. The
patch depends on Impala side change for CDPD-16350.

Testing: lots of plan changes but mainly because the
'_impala_builtins' prefix is dropped from the builtin scalar functions.

Change-Id: I24de64eb7edf8aed49f124542ec4fe5810c7316e