CDPD-19716 HIVE-24031: Infinite planning time on syntactically big queries (#1424) (Stamatis Zampetakis reviewed by Zoltan Haindrich)
1. Drop the defensive copy of children inside ASTNode#getChildren.
2. Protect clients by accidentally modifying the list via an
unmodifiable collection.

Profiling shows the vast majority of time spend on creating defensive
copies of the node expression list inside ASTNode#getChildren.

The method is called extensively from various places in the code
especially those walking over the expression tree so it needs to be
efficient.

Most of the time creating defensive copies is not necessary. For those
cases (if any) that the list needs to be modified clients should perform
a copy themselves.

Change-Id: I96ad96d7432d71e3907cc6c9ab1ee6ef748618b9