CDPD-16813: Partition pruning expressions need normalization.
Impala expects all partition pruning expression to contain the column
as the first parameter in a binary expression (e.g 1 >= my_col must be
converted to my_col <= 1). This commit takes care of the normalization
before the pruning.

Also, because partition pruning does a substitution of the partition data
for the slotref when evaluating, it uncovered another issue. This substitution
causes "resetAnalysisState" to get called and then it recalls "analyze". This
had to be emulated in the Expr subclasses.

Change-Id: Ied4620a6779520aaa93bbeb4efbefb57648c71bd