CDPD-17040: Partition pruner created/triggered more than necessary
Two caches are now in place:

1) At rule level: If the rule is called a second time and the top filter
has not changed, the rule will not be triggered.  If there is no filter
and the rule is called a second time without a filter, the rule will not
be triggered
2) The PrunedPartitionList will be cached for a given table name with
partition expression.  If the same table name with the same expression
is seen again, it will not evaluate the partition expressions a second
time.

In addition to this, an extra HMS call is now made. The Calcite statistics
relies on the HMS partitions to be present. We currently fetch this
once per prune operation.  This can potentially be made more efficient
by doing the fetch immediately after the pruning rule is triggered and
we may want to look into this if this becomes a bottleneck.

On this iteration, a BIG HACK was put in. We need the validWriteIdList
for all tables. However, at planner instantiation time, we don't have
all the tables. A table within a filter will be missed. So we gather
all these right before running the pruning rule. There is hacky code
because there will not always be a transaction open when this is hit.
The correct fix is to make sure an ImpalaPrunedPartitionList is created
within all methods of RelOptHiveTable. CDPD-17313 has been filed to
take care of this, and that will be committed shortly after this one.

Change-Id: I9e441a6c4cebb0fc0be05f3bcca5b17ffe770ef7