CDPD-13549: Allow Impala query options without namespace
This patch allows a user to set up Impala's query options without having
to prepend "impala" to each query option.

In addition, this patch allows a user to set up the query option of
'hive.execution.engine' with the key being 'execution.engine'. When a
user updates the value associated with 'hive.execution.engine', the
value associated with 'execution.engine' would be updated as well and
vice versa.

One thing to note for this patch is that this patch processes a user's
changes to Impala's query options in ImpalaHelper.java instead of in
HiveConf.java. This is because HiveConf.java is currently built in the
first phase of FENG's build procedure and that the Impala's query
options, i.e., TQueryOptions, will not be built until the first phase
of FENG's build is complete. Thus, the changes to Impala's query options
can only be processed in ImpalaHelper.java, when the class of
TQueryOptions has been built. Therefore, a change made by a user to an
Impala's query option without Impala's namespace will not take effect
until the user issues a query for the Impala backend to execute. For
instance, the result for "SET impala.MT_DOP=1" would be either undefined
or its previous value immediately after the execution of "SET MT_DOP=1"
until the user executes a new query. This also implies the value of a
prefixed Impala option could be different from the non-prefixed one. In
such a case, the non-prefixed option takes precedence over the prefixed
one, meaning that the non-prefixed option will be used as the query
option sent to the Impala backend.

Testing:
 - Manually verified in the logs produced by Impala that the query
   options are correctly passed to Impala.
 - Manually verified in beeline that the value associated with
   'execution.engine' reflects the change to that associated with
   'hive.execution.engine' and the other way around.
 - For each query originally in explain_levels.q, added a query
   following the SET statement without Impala's namespace and verified
   that the returned plan is the same even if Impala's namespace is
   omitted.
 - Added new test cases in set_statements.q to demonstrate that the
   patch works as expected.

Change-Id: I0b37685be751365e00112f3df46e9dd6e6112afc