CDPD-17313: Use proper engine rule helper for pruning.
In RelOptHiveTable, we were only using the proper engine
rule helper in computePartitionList.  This commit uses the
proper engine rule helper for all methods at query time.

At MV registry start time, the Hive partition pruner is still
used for the RelOptHiveTable, but when the table is copied into
the session, it will use the correct pruner.

Change-Id: I8c870405ccb14c2411d65d151ff90176187e69a7