DWX-5498: Set status reporting interval in the query context
The status reporting interval is set to 5 secs (matches that
of Impala). This controls how often the backend sends a
status update which includes the profile information to the
coordinator.

Tested the change on a cluster against a TPC-DS sf600
dataset and running a modified version of tpcds q4 which ran
for about 3 minutes.  Without this patch, the query profile
of the inflight query was not getting updated. Only the start
and end profile was registered. With this patch, the profile
gets updated at periodic intervals. The 'Scan progress' also
shows increasing values whereas previously it showed '0%'.

Change-Id: If557901cf7df55e1a8bc11122c3461a7a06e6226