CDPD-17691: Partition pruning not working for non-numeric columns.
Internally, Impala was creating a StringLiteral and it was not matching
up against an "ImpalaStringLiteral" because the classes were different.

It turns out there was no real need to create an ImpalaStringLiteral, so
it is now created as a StringLiteral as well.

The same problem occurred for DateLiteral as well.

While the problem existed for BoolLiterals, the partitions aren't fetching
correctly which will need to be fixed in another Jira.

Change-Id: I3f1a7acb6e6ea43a6de7bda2aecbbb00c726e0b3