CDPD-18288: FAILED: IllegalStateException null when fetch column stats is enabled
Issue was introduced with HIVE-23493 (join rewrite optimization). Filters were not
pushed on top of TS after the optimization was triggered. After this patch,
filters will be immediately on top of the TS operator as it is expected.

Fixed another issue in the optimization that happens when plans are rooted at a
Project followed by a Sort operator.

Change-Id: If9fcb18e8714b41f2ffdb0589629a6cbd4fc8c9d