CDPD-18118: Improve FENG profile timing metrics
This patch adds additional timing events to the FENG profile and accounts
for a gap between the frontend and backend timelines and the time
reported by the client.

The Ping call that is made before statement execution will now receive a
timestamp from the coordinator which it will send back with the
ExecRequest to use as the start time for the backend timeline.

The Impala backend handle is now closed later after the fetch from Hive
has completed so that the backend timeline includes that time as well.
There is a new parameter hive.impala.fetch.early.close that reverts to
the old behavior where the backend handle was close after fetch from
Impala had completed.

Also fixed an issue where AbstractOperatorDesc was throwing many silent
NPEs and an issue where statement that did not fetch rows from the
backend left the backend handle open indefinitely.

Testing: Manually tested with TPC-DS queries
Change-Id: I361fafda8c5023da16e6ee6094220ba3e1dd62cd