CDPD-18117: Store converted Partition object in HMS Cache.
This logic will kick in when the getPartitionsByNames and getTable are
called and the FileMetaData conversion is fetched. The
HiveMetaStoreClientWithLocalCache will use the PartitionCacheHelper object
to help with processing the cache. Within the PartitionCacheHelper, the
partitions are broken up into ones that are found in the cache and ones
that are not.

The new cache object that is stored in the cache is of type
PartitionCacheHelper.CacheValue.  The Result is stored as well as a
convenience map of partition name to the Partition object.

The PartitionCacheHelper will call out to the HMSConverter object if it
exists. If the engine is set to Impala, this will be the ImpalaHMSConverter.
For partitions, the ImpalaHMSConverter will call out to the
ImpalaHMSPartitionConverter object to convert the Partition objects to
HdfsPartition objects.

The ImpalaGetPartitionsByNamesResult class is derived off of the
GetPartitionsByNamesResult class. Since the CacheValue holds a
Result object, it can hold the derived class's object.

Similar logic is used with ImpalaGetTableResult.

To make this work, other small fixes were implemented, including:

- The ImpalaHdfsTable object now fetches the partitions outside of
the constructor.
- The ImpalaHdfsPartition no longer relies on an HdfsTable object.
To avoid this, a couple of methods needed to be overridden in the
HdfsPartition class.

This commit does not contain secrets

Change-Id: Iee4db96bfb2aa36d919b17919930bb06453ee45b