CDPD-18886: Create ImpalaHdfsTable with an ordered list of partitions
The motivation for doing the ordering of partitions is that the
scan ranges are computed in a certain order and that improves the
chances that the Impala backend will assign the scan ranges to
nodes in some deterministic manner.

Tested by dumping the first 5 partitions in the debug log and
running q40 multiple times.

Change-Id: Ib37167f2fc3c1c9e9cdfc9b2032cb0f661a66991