CDPD-15339: Cache the executor membership snapshot
The Impala executor membership snapshot is cached at the
ImpalaSessionManager such that it can be shared across
sessions. The cached entry has a fixed expiration after
which it is reloaded.

Tested through the debugger by running queries in a
session which encounter the cached entry and after the
expiration timeout there's cache miss and cache is reloaded.
Ran all TPC-DS tests via impala-py.test.
Ran queries in 2 concurrent sessions as a sanity check.

Change-Id: I3a5cf803a37a6d858e129118d6e91718ecf8010f