CDPD-19410: Fetch constraints on first creation of RelOptHiveTable.
Fetch the constraints (e.g. foreign key, primary key) when instantiating
the first RelOptHiveTable in SemanticAnalyzer.java

This prevents multiple calls to HiveMetaStoreClient because the constraints
get copied over if there are multiple instances of the table in the query.
This saved roughly 1ms per table instance on my dev environment.

Change-Id: Ic43015f5a4a2fbb3fff52ffe4a00eae7d90f1c10