CDPD-19437: Remove unnecessary initializations on unpartitioned table
On a nonpartitioned table, the ImpalaRulePartitionePruner object was
being instantiated. This class has to extract all the expressions from
the filter that contain partitioned columns. Also, the Impala
HdfsPartitionPruner object is instantiated and needs structures like
TupleDescriptor and Analyzer to work properly.

Because nonpartitioned tables don't need pruning, there is no need to
instantiate some of these objects.

Change-Id: I23a29ee1eb87fbdc94436a9a61587d7137d7ab0c