CDPD-19364: Cache conversion for partition names
The cache conversion feature is now implemented for when the partition
names are fetched from HMS. The cached Impala object is the
ImpalaBasicHdfsTable object.

In order to implement this, the ImpalaBasicHdfsTable had to be made
immutable.

A fix was also made in MetaStoreUtils for when a partition names request is
made when all values need to be fetched.

This commit does not contain secrets

Change-Id: I4a4f7e3878361a219f93726fa91ada2081ca7fbc