CDPD-19511: Buffer size for TSocket is now configurable.
The Thrift TSocket 0.9.3 library has a hardcoded buffer size of 1024
embedded in its code. To get around this, the TCustomSocket
class mimics the TSocket class but allows a configurable buffer size
via the "hive.metastore.thrift.socket.buffer.size" variable.  The default
is 8192.  A value of -1 avoids using a Buffered Stream.

This change also allows TServerSocket to be configured with a customized
buffer size with the class TCustomServerSocket.

Change-Id: I865504913d50c2c1b0e8405b300c2f649c9e2c3b