HIVE-24569: Replace LlapRoutingAppenderPurgePolicy with IdlePurgePolicy and custom appender
Use IdlePurgePolicy to ensure that inactive appenders are always
stopped and removed correctly no matter what happens.

Replace LlapWrappedAppender with LlapRandomAccessFileAppender;
the latter is simpler and unlike its predecessor does not always
rename the file when it is stopped but only when it has received
the specific marker.

Fix Log4jQueryCompleteMarker method that can pose problems if
the marker is used as a filter (not relevant at the moment).

Add tests verifying log file name of query-routing appender;
should contain ".done" suffix when the query is complete.

Remove dead/unused code.