CDPD-21758: Fixed count(*) bug after rebase from master branch
A bug was fixed in the Impala version where the multi aggregate's
substitution match did not have a cloned version of the substitution map.
Because of this, the substitution map was changed by an inner aggInfo layer.

The code in the branch made up for this fact by changing the aggregateExpr if
the substitution happened.  This code is now removed.

Change-Id: I1d2a908a077940008273f27272f274fd37f45bc1