CDPD-21074, CDPD-20772, CDPD-20884: Fix wrong results for query with coalesce and outer join
Ensure that a join's output substitution map is populated with the
the correct mapping when a coalesce() is wrapped within an
IfTupleIsNull predicate. If the execution mode is HIVE_IN_TEST
then a local implementation of the null wrapping check is used.
Otherwise, Impala's implementation is used which indirectly does
function evaluation.

Change-Id: I52733d944c27549d3cd5c9c811bbf967dc969a26