CDPD-21674: exclude unnecessary dependencies from lib
Previous commit on CDPD-21674 excludes just the
dependencies from hive-service module. However,
the hive-packaging module includes all the transitive
dependencies which causes the problematic jars
like hibernate-core to be included in lib directory.

This change moves the pac4j dependency and its
exclusions to the top level dependencyManagement
instead of having it as a dependency in hive-service
module. Additionally, it excludes pac4j from
other modules which have hive-service as a dependency
(eg. hive-jdbc, hive-cli) so that we don't leak it
outside hive-service.

Testing:
1. Ran unit tests for SAML auth.
2. Verified that hibernate*.jar are not present
in lib directory anymore.

Change-Id: I6a08534c437733ee8d189ac99d5ec1d6d9324adc