CDPD-21923: New commit override methods needed in HiveMetaStoreClientWithLocalCache
There were new commit methods that were added to HiveMetaStoreClient after the rebase.
The existing commit methods in HiveMetaStoreClientWithLocalCache may contain stale
data for a transaction id. This can happen when caching is done for a table
within the transaction before the insert into the table. In this case, The existing
commit methods invalidate any data that was cached during the transaction.

The new commit methods need to invalidate the data during the transactions as well.

Change-Id: I496c4e20888e529bfd5be7ef7ad0f5824182fea3