DWX-6609: Fix for browse action is not supported error
It is possible that the even when Desktop API is
supported the BROWSE action is not. This would cause
a error when driver is attempting to open a browser.

This fix adds additional condition to make sure that
BROWSE action is supported and if it is not, it fallsback
to using OS commands to open the browser.

Testing:
This issue was discovered on my Ubuntu laptop when using
DBeaver. After this fix, driver was able to open the
browser and complete the SSO flow.

Change-Id: I3863cc3056b2024f24c544721ff26d725e41ae31