CDPD-22308: Modify a test in expr_return_types.q
One of the tests in expr_return_types.q was using the LAG
window function. When run through maven it fails because we
call Impala's standardize() method which internally tries to
load libfesupport.so which is not currently supported via
the maven tests.

Since the main purpose of the test is to check planning for
window functions, I modified it to use MAX window function
with a very similar plan.

Change-Id: I1f92a0e7326ca0dbec62f53e9d63da54d6e1452a