CDPD-21269: Implement CardinalityCheckRel RelNode.
The CardinalityCheckRel RelNode will replace the sq_count_check built-in function
in Impala.  The functionality is slightly different.

The sq_count_check function checks that the value coming in is 0 or 1. This is done
in a Filter RelNode.

The CardinalityCheckRel RelNode checks that the number of rows entering the node is
zero or one. The node also requires the underlying node to only pass in at most 2 rows.

Change-Id: Id15aaf18475feb190b35c77f5f9c72d72b6c189a