HIVE-24817: A "not in" clause returns incorrect data when there is coercion
When the query has a where clause that has an integer column checking against
being "not in" a decimal column, the decimal column is being changed to null,
causing incorrect results.

This is a sample query of a failure:
select count from my_tbl where int_col not in (355.8);

Since the int_col can never be 355.8, one would expect all the rows to be
returned, but it is changing the 355.8 into a null value causing no rows
to be returned.

The fix is that when the value is attempted to be interpreted and it fails,
the original type should be used rather than null.