CDPD-20870: IN syntax causing impalad server to crash
There were two issues:

1) The IN Calcite function can be created by rules while optimizing. The
correct engine sematics need to be used when this happens.

2) When the IN function contains decimal parameters, it needs
all the parameters to have the same precision and scale. So some
special case logic had to be implemented.

Also made a change so that all chars are compatible with each other.
At parsing compile time, this is the right thing to do because a
char(6) can be compared to a char(5).  The runtime (or optimization)
code should deal with the fact that a char(6) is not equal to a char(5).

Change-Id: I58895000dcdd8d187baceeb39065d834c1601302