CDPD-20764: Correlated fields not detected if subquery has where and group by clauses.
In order to detect correlated queries, a traversal was being done through the
RelNodes in the subqueries.  The traversal stopped when it hit the first HiveFilter
and checked for correlated fields in the condition. This didn't work when there
was both a having clause and a where clause since both clauses are HiveFilter
RelNodes.

The fix is to traverse through all HiveProject, HiveAggregate, and HiveFilter
clauses. This is a slight expansion over what was previously done. The code
still does not go through other RelNodes to keep in line with the code that
was there previously.

Change-Id: I3b021b1c67188e9496b3f99dd5f1d9022b9d0ace