CDPD-20799: Fetching wrong signature when function has variable args.
Fixed an issue with a function containing variable arguments where the
wrong signature was fetched.

Specifically, the coalesce function was being called with arguments
(DATE, TIMESTAMP).  The candidate function picked was coalesce(DATE)
because that matched the first parameter.  However, all parameters need
to be inspected when there are variable arguments.

Change-Id: I77cc046ab6fbf28a88351132ee07e333fd990bd4