CDPD-23225: Support REFRESH sql statement for external tables in FENG.
Refresh command is submitted to the Impala engine as a SQL statement for execution in synchronous mode. The semantic checks performed on hive side are:
     1. Only external tables are allowed (and not views).
     2. Partition spec (if specified) should be full and not partial
The REFRESH metadata authorization is deferred to Impala since we expect Impala coordinators to be configured with Ranger.

The sql syntax of REFRESH command is same as that in Impala.

Change-Id: If71f41e75655f657e657f143979ec9569684f03d