CDPD-24011: The functions date_add, date_sub returning incorrect results.
The date_add and date_sub functions were being processed as generic queries.
They need to be treated as "+" operations so that specific date intervals
like "years, "days", etc... are processed with existing time interval logic
handled by the "+" and "-" operators. This only needs to happen when the
second parameter is an interval type.  When it is an int type, it can be
processed like other "<time_interval>_add/sub" functions.

The fix for CDPD-20416 is included as well. The month*interval function should
be used when using syntax like "interval 3 months".

Change-Id: Ie4caaa312ad0149ab6306a39e215186018901c01