CDPD-24017: Handle overflow for year interval operations.
When using Hive semantics with a high value such as MAX_INT, the value was
overflowing causing a small negative integer to be added to the timestamp
column.

Fixed this by retrieving  an intValue and multiplying by 12L (a long value
for months).

Impala doesn't handle a years value this big anyway, but having a small
negative value added to the timestamp needs to be avoided.

Change-Id: I9364e5ccbd87103f6afdfe94030dcb27951648e5