CDPD-24255: Need to use common decimal type for case functions
The change to use common types was made recently for the IN clause.
It is now applied to the case clause as well.

The logic has also been changed for the return type of the case clause.
Previously the code was using Calcite's mechanism to infer the return
type, but now it uses the Impala semantics.

Change-Id: I3e3146da5d585954aae88e56d887b6184fa470a9