DWX-7104: Added extra HMS call for materialized views to fetch the Table object.
At initialization time, the Table objects are fetched for all materialized views.
However, the accessType attribute is not included in the Table object. This is
because the processor capabilities is not included in the fetch call.

The optimal solution would be to include the capabilities (XXX: fill in jira here).
A quick fix to refetch the Table is included in this commit until the thrift
file upgrade can be done.

Change-Id: Icf8841939ea1196fde346d2aa8c2461ed5c24315