CDPD-24446: Support limit pushdown for analytic sort
This patch ports the changes for doing limit pushdown into
analytic sort (under qualifying conditions) from Impala's FE.
The reason for porting instead of directly using Impala's
functions is that there are a few customizations needed for
the FENG plan. The comments in the code highlight those
differences.

The optimization mainly affects TPC-DS q67. Also ran the
end-to-end Impala tests for this optimization.

Change-Id: I7cef579b1c3c423273b54cebc2acd6d11fa6aa3f