HIVE-24957: Wrong results when subquery has COALESCE in correlation predicate
1. Add plan transformations before starting the core RelDecorrelator logic
to bring the plan into an equivalent but more convenient form that can be
decorrelated into more efficient and correct plans.

2. Adapt HiveRelDecorrelator#decorrelateInputWithValueGenerator to avoid
creating value generator for already satisfied correlations present in
the input.

3. Based on the changes above many plans with subqueries become more
efficient since the value generator is no longer necessary and it is dropped.

4. Add subquery_complex_correlation_predicates.q which includes queries
generating wrong results without the new transformations.

5. Add CBO plans in few queries since they are easier to read and reason
about correctness and efficiency.