DWX-7217: Only set GetPartitionsByNames.GetFileMetadata when true.
This fix is to temporarily work around a bug on the server side.
On the server side, there is a bug that checks isSetGetFileMetadata()
rather than isGetFileMetadata().  So if the value is explicitly set to
false, the server side currently thinks it's true.

The code will be fixed on the server side, but the change here is still
correct code and it's being made to expediate the testing.

Change-Id: I6c8cf5b9d98ed03fa64fd939f60810d238c63a9b