HIVE-25296: Replace parquet-hadoop-bundle dependency with the actual parquet modules
The parquet-hadoop-bundle is not a real dependency but a mere packaging
of three parquet modules to create an uber jar. The Parquet community
created this artificial module on demand by HIVE-5783 but the
benefits if any are unclear.

On the contrary using the uber dependency has some drawbacks:
-Parquet souce code cannot be attached easily in IDEs which makes
debugging sessions cumbersome.
-Finding concrete dependencies with Parquet is not possible
just by inspecting the pom files.
-Extra maintenance cost for the Parquet community adding
additional verification steps during a release.

For the reasons above the uber dependency is replaced with concrete
dependencies to the respective modules:
* parquet-common
* parquet-column
* parquet-hadoop

No need to exclude ord.codehaus.jackson from parquet in the shaded
hive-jdbc jar cause its not a compile dependency in the afforementioned
Parquet modules (at the current version).