CDPD-26596: Fetch validWriteIdList for tables before table request.
Added code to ensure the validWriteIdList for all tables are fetched
before any table request is done.

The tables are now gathered at parsing time within FromClauseParser.g.
One of the initial steps while parsing is to make an HMS call to fetch
the validWriteIdList for all parsed tables. This is done within a newly
added class called CacheTableHelper. The only purpose of this class is
to populate the query HMS cache for future calls to HMS. The methods to
this class do not contain any return values.

One complication to sending the request up front is when the query contains
views. The views will have underlying physical tables that need fetching but
we won't know these tables until after doing an HMS call. This is handled
by using the underlying tables of the most recent time a query used this view.
The idea here is that views don't change very often, so we will most likely
be able to cache these tables by guessing that these tables are still in the
view. On the chance that the view did change, we may be doing a fetch of
validWriteIdLists on tables no longer used or we may miss tables added on
the new definition of the view. If that happens, the only downside is an
extra HMS call as we get the updated information of the view (which cannot
be optimized). The view will then be updated so that the next query will
be optimized.

Because this view information is cached at a server level, a cap of 10000
tables to track was used. Also, on the low chance there is a problem with
memory consumption on this feature, it can be turned off via the
HIVE_OPTIMIZE_VIEW_CACHE_ENABLED conf variable.

Change-Id: I0037a7285de6a09770cda626c44a71c797323ad2