HIVE-25173: hive-pre-upgrade build fails due to missing dependency on pentaho-aggdesigner-algorithm (Masatake Iwasaki, reviewed by Stamatis Zampetakis)
org.pentaho:pentaho-aggdesigner-algorithm dependency comes transitively from calcite-1.10.0
and was hosted in conjars.org repo. The latter has been retired recently and the dependency
is not present in maven central so the build fails when the dependency is not present in
the local maven repo.

The devs can workaround the problem by defining another maven repo (e.g., in settings.xml)
where the dependency is present but since it is not mandatory for hive-pre-upgrade we can
safely exclude it.

Closes #2326