HIVE-25403: Fixing from_unixtime() in GenericUDFFromUnixTime class so as to consider leap seconds
As per the official doc, the func from_unixtime() takes only unix time in seconds. So removed IntObjectInspector, and kept only LongObjectInspector

What changes were proposed in this pull request?
i) Fixed the code in the evaluate() in GenericUDFFromUnixTime to consider leap seconds.
ii) Refactor GenericUDFFromUnixTime class to make code much more clean and readable.

Why are the changes needed?
SELECT from_unixtime(unix_timestamp(cast('1400-01-01' as date)));
was giving wrong results, because the from_unixtime() function was not considering leap seconds while representing Timestamp. So it was required to change the epoch time in the
from_unixtime() to ZonedDateTime.
The required conversion was done with the help of the Instant class, which represents a moment given the epoch time.

Does this PR introduce any user-facing change?
No

How was this patch tested?
Added relevant qtests in udf5.q
Added unit tests in TestGenericUDFFromUnixTime.java