HIVE-25553: Support Map data-type natively in Arrow format
This covers the following sub-tasks as well:
HIVE-25554: Upgrade arrow version to 0.15
HIVE-25555: ArrowColumnarBatchSerDe should store map natively instead of converting to list

What changes were proposed in this pull request?
a. Upgrading arrow version to version 0.15.0 (where map data-type is supported)
b. Modifying ArrowColumnarBatchSerDe and corresponding Serializer/Deserializer to not use list as a workaround for map and use the arrow map data-type instead
c. Taking care of creating non-nullable struct and non-nullable key type for the map data-type in ArrowColumnarBatchSerDe

Why are the changes needed?
Currently ArrowColumnarBatchSerDe converts map datatype as a list of structs data-type (where struct is containing the key-value pair of the map).
This causes issues when reading Map datatype using llap-ext-client as it reads a list of structs instead.
HiveWarehouseConnector which uses the llap-ext-client throws exception when the schema (containing Map data type) is different from actual data (list of structs).
This change includes the fix for this issue.

Does this PR introduce any user-facing change?
No

How was this patch tested?
Enabled back the Arrow specific tests in Hive code