HIVE-25625: Drop TestCompareCliDriver and related code from tests
The driver has been introduced back in 2015 aiming to run queries with
vectorization on/off and comparing the results. However it didn't
receive much attention since then and currently only two queries are run
with this driver. The majority of tests aiming to ensure vectorization
works correctly use the TestMiniLlapLocalCliDriver and run a query
twice switching on/off the necessary properties.

Summing up having the CoreCompareCliDriver in the repo leads to extra
code maintenance cost without much benefit.