Always display SQLLine's standard out & error in debug logs
1. Drop LOG.isDebugEnabled() checks and use of NullOutputStream.
This is test code and almost always we run with debug logs enabled so
optimizing to use NullOutputStream does not make much sense. Moreover,
the scripts are not huge so optimizing the code for this is premature.

2. Include standard error in the debug logs
When a script fails the only way to know what happened is to look in the
standard error.