Align order of obtaining JDBC password in CalcitePlanner with JdbcStorageConfigManager
Retain the previous functionality where we first check PASSWORD,
and then KEYSTORE, and check URI at the end as it is done also in
JdbcStorageConfigManager.

Remove 'skipping CBO' from warning message when password is null. The
phrase is not necessarily true. Null is an accepted value when the JDBC
endpoint does not require a password.